+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 	Placeholder for the next version (at the beginning of the line):
 	## __WORK IN PROGRESS__
 -->
+
+## **WORK IN PROGRESS**
+
+-   [#104](https://github.com/MiSchroe/klf-200-api/issues/104) Add LimitationOriginator and LimitationTime properites.
+
 ## 3.4.0 (2024-02-08)
 
 -   [#98](https://github.com/MiSchroe/klf-200-api/issues/98) Add Products.requestStatusAsync to get the latest value from one or more products.